refactor(api): extract auth-core + auth + users domains (#35)
Decomposes the auth/users slice of api/handlers.ts into api/handlers/auth-core.ts
(JWT secret singleton + password/token primitives + getAuth/requireOwner guards),
api/handlers/auth.ts (thin login/me/status route handlers), and
api/handlers/users.ts (users CRUD handlers). Pure move, zero behavior change.

api/handlers.ts becomes further thinned as a hybrid shim: explicit named
re-exports for the moved public symbols, still-inline for the remaining
domains (languages, pages, menus, redirects, configs, media, global-blocks,
cache, backup-routes). The JWT secret singleton's module-load-time console.warn
side effect and fail-closed production behavior are preserved byte-for-byte
(security-sensitive, GHSA-mfcg).
